+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "sanita-ticket-js",
-  "version": "1.5.1",
+  "version": "1.6.0",
   "description": "Electron app to read 2003 mdb files and create a purchase ticket.",
   "keywords": [
     "electron",
