+++ release/app/package.json
@@ -1,6 +1,6 @@
 {
   "name": "sanita-ticket",
-  "version": "1.5.1",
+  "version": "1.6.0",
   "description": "Budget & and tickt system for sanita",
   "license": "MIT",
   "author": {
