+++ src/renderer/components/ViewTicketModal.tsx
@@ -2,8 +2,8 @@ import { FC } from 'react';
 import { Modal, ModalClose, ModalDialog, DialogTitle, Stack, Button, Box } from '@mui/joy';
 import { Print } from '@mui/icons-material';
 import { Ticket } from './Ticket';
-import { IHistoryItem } from '../../types';
 import { useTicketSummary } from '../hooks/useTicketSummary';
+import { IHistoryItem } from '../../types';
 
 export interface IViewTicketModalProps {
   ticket: IHistoryItem;
