+++ src/types/auth.ts
@@ -1,0 +1,6 @@
+export interface IUser {
+  id: string;
+  email: string;
+  role: string;
+  isAdmin: boolean;
+}
