+++ src/types/index.ts
@@ -1,0 +1,4 @@
+export * from './auth';
+export * from './tickets';
+export * from './history';
+export * from './products';
